refactor: extract static-NAT compile into its own file (#4846)
Move compileNATStatic and its static-only helpers (natStaticPrefixInfo,
isStaticBlockPair, applyStaticNATFromScope, staticNATMappedPortFromKeys,
staticNATRoutingInstanceFromKeys, resolveStaticNATThenPrefixName,
resolveStaticNATThenPrefixNames) out of compiler_nat.go into a new
compiler_nat_static.go in the same package.

Pure code motion: bodies moved byte-for-byte, no signature or logic
change. The new file imports only strconv and strings. What remains in
compiler_nat.go is now just the source-NAT compile concern.
